+++ src/include/distributed/metadata_cache.h
@@ -40,7 +40,7 @@ typedef struct
 	/* pg_dist_partition metadata for this table */
 	char *partitionKeyString;
 	char partitionMethod;
-	uint64 colocationId;
+	uint32 colocationId;
 	char replicationModel;
 
 	/* pg_dist_shard metadata (variable-length ShardInterval array) for this table */
